moved some code around, hid some intellij files
