+++ .gitignore
@@ -10,3 +10,9 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+# IntelliJ project files
+*.iml
+.idea/
+out/
+src/META-INF/MANIFEST.MF
