--- Time.java
@@ -1,13 +1,0 @@
-public class Time extends Command {
-	public Time() {
-		startorall = false;
-        trigger = ".time";
-		name = "Time";
-		description = "Displays time for an example";
-	}
-	
-	public String Execute (String channel, String sender, String login, String hostname, String message) {
-		String time = new java.util.Date().toString();
-		return sender + ": The time is now " + time;
-	}
-}
